General Cleanup and minor additions
FIXES
- Fixed ZeldaPlayerHealth bug that was adding +2 Health to new players this should no longer happen and the player should spawn with 20 Health.
- Hero Bow now works again due to error with renaming "Lightning_Arrow"
 -> "Shock_Arrow" and not properly updated the tag bow_ammo.
 - Light Arrows and Ancient Arrows now emit particles when colliding with anything
 - Gloom Container consume duration is returned to former value

CLEANUP
- ZeldaRedstone Block has been implemented that contains the Push, Pull, and Press methods
- Eye Switches,Pressure Switches, and Shock Switches now properly extend each other and make use of new ZRB. Code has been generally cleaned up.
- Spike Blocks properly interact with redstone blocks
- Spike Blocks now use Cactus damage source instead of custom one due to bug with knockback. Might revise later on.

CHANGES
- Hammers now create an AOE attack with matching effect upon right-clicking. Their descriptions have been adjusted to account for this change.
- Reforged Master Sword now properly extends the Master Sword and no longer uses right-click to shoot beams.
- Blast Mask has been rewritten to cleanup code. It can now be used to Bomb Jump if used whilst blocking!

- SwordBeam2 -> SwordBeam_Evil

+++ src/main/java/com/kamth/zeldamod/item/items/movement/ScaleItem.java
@@ -47,7 +47,7 @@ public class ScaleItem extends Item {
             f1 *= f5 / f4;
             f2 *= f5 / f4;
             f3 *= f5 / f4;
-            player.push((double) f1, (double) f2, (double) f3);
+            player.push(f1, f2, f3);
             player.startAutoSpinAttack(20);
             player.getCooldowns().addCooldown(itemstack.getItem(), 40);
             return InteractionResultHolder.success(itemstack);
